requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
numpy==1.24.3
tensorflow==2.15.0
google-generativeai==0.3.2
Pillow==10.1.0
gunicorn==21.2.0
